Add files via upload
A design pattern is a reusable solution to a common problem that occurs in software design and
development. It is a proven approach to solving a specific design or architectural issue in a
consistent and efficient manner. Design patterns capture best practices and provide a way to
communicate and share effective solutions among developers
Following are the design patterns
 Singleton design pattern
 Facade design pattern
 Builder design pattern
 Iterator design pattern